feedback-ui/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
httpx>=0.27.0
python-dotenv>=1.0.0
jinja2>=3.1.0
python-multipart>=0.0.6
xai-sdk>=0.1.0
pandas>=2.0.0
